[fix] Rename some filename from asu_* to kv_client_*
