+++ kv_semantics/kv_test/include/bench_runner.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "asu_client_runner.h"
+#include "kv_client_runner.h"
 #include "kv_test_types.h"
 
 namespace kv::bench {
