+++ kv_semantics/kv_test/include/kv_test_app.h
@@ -1,12 +1,12 @@
 #pragma once
 
 #include "arg_parser.h"
-#include "asu_client_runner.h"
 #include "bench_runner.h"
 #include "buffer_allocator.h"
 #include "consistency_checker.h"
 #include "hcomm_config_adapter.h"
 #include "key_value_generator.h"
+#include "kv_client_runner.h"
 #include "kv_test_config_loader.h"
 #include "result_writer.h"
