+++ kv_semantics/kv_test/src/kv_runtime_proxy.cc
@@ -1,4 +1,3 @@
-#include "asu_runtime_proxy.h"
 #include <cstdlib>
 #include <dlfcn.h>
 #include <filesystem>
@@ -6,6 +5,7 @@
 #include <unistd.h>
 #include <utility>
 #include <vector>
+#include "kv_runtime_proxy.h"
 
 namespace kv::bench {
